requirements.txt
numpy
pandas
matplotlib
requests
yfinance
alpha_vantage
fredapi
python-dotenv
pandas-datareader
openai
